PROYECTO/BLOQUE_3/backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
google-generativeai
groq
